refactor: move initial plugins logic to useAppStartup
